Improve tooltip/page/route components code
