+++ apps/extension/src/hooks/useCraftableItemPage.ts
@@ -1,0 +1,14 @@
+import * as React from 'react';
+import { useItemsFetcher } from './useItemsFetcher';
+
+export function useCraftableItemPage(reagentItemIds: number[]) {
+  const id = React.useId();
+  const [craftAmount, setCraftAmount] = React.useState(1);
+  const items = useItemsFetcher(id, reagentItemIds);
+
+  return {
+    craftAmount,
+    setCraftAmount,
+    items,
+  };
+}
